requirements.txt
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pycparser==2.20
pymongo==3.11.0
python-dotenv==0.14.0
six==1.15.0
waitress==1.4.4
Werkzeug==1.0.1
WTForms==2.3.3
